feat: Change precision to 2 decimal places. Allow non-lockup accounts watch for rewards (#8)
* Change precision to 2 decimal places to allow non-lockup non-whales accounts to watch for rewards

* Bump version to 0.4.0 add rust-toolchain

+++ src/utils/human.rs
@@ -1,5 +1,5 @@
-pub(crate) fn human(yocto: u128) -> u128 {
-    yocto / 1000000000000000000000000
+pub(crate) fn human(yocto: u128) -> f64 {
+    yocto as f64 / 1000000000000000000000000_f64
 }
 
 pub(crate) fn current_position_in_epoch(start: u64, current: u64) -> u64 {
